Fixes. Added the cGui stuff to msvc project. Pelya, please fix your code, and this goes to everyone: TEST-COMPILE BEFORE UPLOADING. Comment with more stuff included in one of the files, cause i made a (very)dirty fix.
All variables starting with VT_ should be avoided, as those are already decleared in wtypex.h. (Renamed vt_ to svt_)

Stick to "it" for loop variables when you use iterators, especially if i has been function-declared as unsigned int a bit up ..... Also, this is standard already, stick to it.

git-svn-id: https://openlierox.svn.sourceforge.net/svnroot/openlierox@1065 34602234-ff1f-0410-a465-ea8f3b77ab7f